fix(kyc): cleanup and correctness fixes from code review
- fix support email: support@superteamearn.com → support@superteam.fun in all KYC error messages
- fix HTTP status: "Not allowed" responses now return 403 instead of 200 (submission + grant verify-completion and grant verify-poa)
- remove dead code: getKycRejectionReason helper in both KYC modals — checked for a KYC_REJECTED error shape the backend never returns
- extract shared helper: createSumsubAccessToken util used by both access-token and poa-access-token endpoints instead of duplicating the same axios call
