+++ src/pages/api/grant-application/kyc/verify-poa.ts
@@ -39,7 +39,7 @@ const handler = async (req: NextApiRequestWithUser, res: NextApiResponse) => {
       grantApplication.user.isKYCVerified;
 
     if (!isAllowed) {
-      return res.status(200).json({ message: 'Not allowed' });
+      return res.status(403).json({ message: 'Not allowed' });
     }
 
     const secretKey = process.env.SUMSUB_SECRET_KEY;
