+++ src/pages/api/sumsub/access-token.ts
@@ -1,40 +1,8 @@
-import axios from 'axios';
 import type { NextApiResponse } from 'next';
 
 import { type NextApiRequestWithUser } from '@/features/auth/types';
 import { withAuth } from '@/features/auth/utils/withAuth';
-import { SUMSUB_BASE_URL } from '@/features/kyc/constants/SUMSUB_BASE_URL';
-import { type SumSubBaseResponse } from '@/features/kyc/types/SumSubBaseResponse';
-import { createSumSubHeaders } from '@/features/kyc/utils/createSumSubHeaders';
-import { handleSumSubError } from '@/features/kyc/utils/handleSumSubError';
-
-const createSumsubToken = async (
-  userId: string,
-  secretKey: string,
-  appToken: string,
-) => {
-  const url = '/resources/accessTokens/sdk';
-  const method = 'POST';
-  const body = JSON.stringify({
-    ttlInSecs: 600,
-    userId,
-    levelName: process.env.SUMSUB_LEVEL_NAME,
-  });
-
-  const headers = createSumSubHeaders(method, url, body, secretKey, appToken);
-
-  try {
-    const response = await axios.post<SumSubBaseResponse>(
-      `${SUMSUB_BASE_URL}${url}`,
-      { ttlInSecs: 600, userId, levelName: process.env.SUMSUB_LEVEL_NAME },
-      { headers },
-    );
-    return response.data;
-  } catch (error) {
-    handleSumSubError(error);
-    throw error;
-  }
-};
+import { createSumsubAccessToken } from '@/features/kyc/utils/createSumsubAccessToken';
 
 const handler = async (req: NextApiRequestWithUser, res: NextApiResponse) => {
   const userId = req.userId;
@@ -42,12 +10,18 @@ const handler = async (req: NextApiRequestWithUser, res: NextApiResponse) => {
   try {
     const secretKey = process.env.SUMSUB_SECRET_KEY;
     const appToken = process.env.SUMSUB_API_KEY;
+    const levelName = process.env.SUMSUB_LEVEL_NAME;
 
-    if (!secretKey || !appToken || !userId) {
+    if (!secretKey || !appToken || !userId || !levelName) {
       return res.status(500).json({ message: 'Missing environment variables' });
     }
 
-    const result = await createSumsubToken(userId, secretKey, appToken);
+    const result = await createSumsubAccessToken(
+      userId,
+      levelName,
+      secretKey,
+      appToken,
+    );
 
     return res.status(200).json(result);
   } catch (error) {
